requirements.txt
flask==2.0.1
python-docx==0.8.11
pandas==1.3.3
torch==1.9.0
transformers==4.11.3
pytest==6.2.5
pytest-cov==2.12.1
gunicorn==20.1.0
flake8==3.9.2
black==21.7b0